+++ packages/vseed/src/pipeline/advanced/chart/pipeline/column.ts
@@ -17,9 +17,9 @@ import {
 
 export const columnAdvancedPipeline: AdvancedPipeline = [
   initAdvancedVSeed,
+  encodingForColumn,
   buildMeasures,
   autoDimensions,
-  encodingForColumn,
   pivotAdapter([reshapeWithEncoding], [pivotReshapeWithEncoding]),
   sortXBandAxis,
   sortLegend,